finalizing basic tests & adding stubs
